Agentic AI Q4/Class Code/Class-5 (18 July 2025 )/pyproject.toml
[project]
name = "class-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai==0.3.2",
    "pymupdf==1.23.22",
    "pypdf2>=3.0.1",
    "python-dotenv==1.0.1",
]
